# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cairn API
  version: 0.1.0
  description: Versioned API shared by Cairn desktop, PWA, and future mobile clients.
  license:
    name: GPL-3.0-only
servers:
  - url: http://127.0.0.1:7381
paths:
...
  /api/v1/folders:
    get:
      operationId: listFolders
      tags: [subscriptions]
      responses:
        "200":
          description: Folder hierarchy source data.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Folder"
    post:
      operationId: createFolder
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "201":
          description: Folder created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
  /api/v1/folders/{folderId}:
    parameters:
      - $ref: "#/components/parameters/FolderId"
    patch:
      operationId: updateFolder
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
      responses:
        "200":
          description: Folder updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "409":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteFolder
      tags: [subscriptions]
      responses:
        "204":
          description: Folder deleted.
        "404":
          $ref: "#/components/responses/Problem"
...
  /api/v1/entries/{entryId}/tags:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    put:
      operationId: setEntryTags
      tags: [entries, tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_ids]
              properties:
                tag_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Entry tags replaced.
          content:
            application/json:
              schema:
                type: object
                required: [tag_ids]
                properties:
                  tag_ids:
                    type: array
                    items:
                      type: string
                      format: uuid
  /api/v1/tags:
    get:
      operationId: listTags
      tags: [tags]
      responses:
        "200":
          description: Profile tags.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
    post:
      operationId: createTag
      tags: [tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: Tag created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /api/v1/tags/{tagId}:
    delete:
      operationId: deleteTag
      tags: [tags]
      parameters:
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Tag deleted.
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/rules:
    get:
      operationId: listRules
      tags: [rules]
      responses:
        "200":
          description: Automation rules.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleList"
    post:
      operationId: createRule
      tags: [rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "201":
          description: Rule created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /api/v1/rules/{ruleId}:
    delete:
      operationId: deleteRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "204":
          description: Rule deleted.
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/saved-filters:
    get:
      operationId: listSavedFilters
      tags: [filters]
      responses:
        "200":
          description: Saved timeline filters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedFilterList"
    post:
      operationId: createSavedFilter
      tags: [filters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavedFilterRequest"
      responses:
        "201":
          description: Saved filter created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedFilter"
  /api/v1/saved-filters/{filterId}:
    delete:
      operationId: deleteSavedFilter
      tags: [filters]
      parameters:
        - $ref: "#/components/parameters/FilterId"
      responses:
        "204":
          description: Saved filter deleted.
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: CairnDeviceToken
  parameters:
...
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    FolderId:
      name: folderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TagId:
      name: tagId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RuleId:
      name: ruleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FilterId:
      name: filterId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Folder:
      type: object
      required: [id, name, position, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: [string, "null"]
          format: uuid
        name:
          type: string
        position:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateFolderRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        parent_id:
          type: [string, "null"]
          format: uuid
    UpdateFolderRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        parent_id:
          type: [string, "null"]
          format: uuid
        position:
          type: integer
    Tag:
      type: object
      required: [id, name, position, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: [string, "null"]
        position:
          type: integer
        created_at:
          type: string
          format: date-time
    TagList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
    CreateTagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        color:
          type: [string, "null"]
    Rule:
      type: object
      required: [id, name, enabled, priority, conditions, actions, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        conditions:
          type: object
          additionalProperties: true
        actions:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RuleList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Rule"
    CreateRuleRequest:
      type: object
      required: [name, conditions, actions]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        enabled:
          type: boolean
          default: true
        priority:
          type: integer
        conditions:
          type: object
          additionalProperties: true
        actions:
          type: object
          additionalProperties: true
    SavedFilter:
      type: object
      required: [id, name, query, position, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        query:
          type: object
          additionalProperties: true
        position:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SavedFilterList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SavedFilter"
    CreateSavedFilterRequest:
      type: object
      required: [name, query]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        query:
          type: object
          additionalProperties: true
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        request_id:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
security:
  - deviceToken: []